services/anomaly-engine/requirements.txt
kafka-python==2.0.2
psycopg[binary]==3.2.1
redis==5.0.8
